+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+[Full Changelog](https://github.com/netzmacht/release-notifier/compare/0.4.1...master)
+
+### Fixed
+
+- Sort packagist releases ascending
+
 ## 0.4.1 - 2019-06-14
 
 [Full Changelog](https://github.com/netzmacht/release-notifier/compare/0.4.0...0.4.1)